backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
SQLAlchemy==2.0.36
aiomysql==0.2.0
pydantic==2.9.2
python-dotenv==1.0.1
